[#40] Chore :  Application 전용 Convention Plugin 추가
