+++ build-logic/convention/src/main/kotlin/ConfigFlavorPlugin.kt
@@ -1,4 +1,5 @@
 import com.android.build.api.dsl.CommonExtension
+import com.dhc.buildlogic.util.configureFlavor
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.findByType
@@ -10,27 +11,4 @@ class ConfigFlavorPlugin : Plugin<Project> {
             target.configureFlavor(extension)
         }
     }
-}
-
-internal fun Project.configureFlavor(
-    commonExtension: CommonExtension<*, *, *, *>,
-) {
-    with(commonExtension) {
-        productFlavors {
-            Flavor.values().forEach { flavor ->
-                if (flavorDimensions.contains(flavor.dimen).not()) {
-                    flavorDimensions += flavor.dimen
-                }
-
-                create(flavor.flavor) {
-                    this.dimension = flavor.dimen
-                }
-            }
-        }
-    }
-}
-
-private enum class Flavor(val flavor: String, val dimen: String) {
-    DEV("dev", "server"),
-    PROD("real", "server")
 }
